CPU emulator and small instruction set
Rust's match statement proves to be terse compared to the verbose
c/c++ switch statement; they are alternatives but one is a pain to work
with especially the default fallthrough mechanism that sometimes is a
bug that is a pain in the a** to find. As for the emulator, I'm yet to
write a small program to be run by the emulator, task for another day,
fibonacci seems like the perfect one.