Operator Overloading and intro to Module system
Rust also happens to support operator overloading without weird syntax
like in C++. The trait system has proven to be quite capable and if I'm
right, traits offer virtual functions. I rust knows the exact function
called, it will execute it imediately, if not, it must be in the vtable.