Some standard rust traits implemented for custom Shape enum
Turns out, matching enums with named fields have to bind to the same
names for use, this is brilliant, it reduces errors and stup*d mistakes
we make in development.
Though, the io seems to be quite interesting. If I'm not mistaken, one
has to lock the whole programs IO to do either in or out operations, it
does achieve synchronisation but at the cost of the syntax, one cannot
print/read from terminal if someone else is holding the lock to
read/print, more on this later.