Error handling in rust
Inplace of throw and raise, rust has returns, pretty neat if you ask me.
rust provides two types of error propagation, 1) panic the hell out of
it, or 2) package it as a Result::Err(YourErrorT) and return to the
caller for them to handle. It is recomended to use Result if the error
is recoverable, since this allows the caller to try something else or
fix the problem, while as for panic, this is where you terminate the
application telling the developer there's a bug in there code and the
error cannot be ignored since it would possibly leave the application in
a broken/dangling state.