Intro to rust generics
Being familiar to c++ templates, rust generics were not that hard on me
except for the lifetime inclusion, they really do make things a bit
cumbersome. But so far my intuition on lifetimes is improving; lifetimes
prevent a reference from going back in scope in reference to the
refererred while the borrow checker ensure the values are used in ranges
where they are okay, that is, from initialization to moved/dropped
after/before which it is invalid.