Testing in Rust
One of the beautiful aspects of rust is testing. Unlike languages like C/C++ that
not only lack standard build systems but also testing specifications
make using them a nightmare as you are left alone in the dark to setup
testing and buildtools to use the languages. Rust, however, comes with
pretty standard tool, cargo, that is just extensible and very helpful.
Testing in rust is amazing and in this commit, we have a small program
that essentially does nothing but explore what rust has to offer for
testing. In the creation, I had a pretty nasty bug, a logical one, and
it was well hidden but exposed via testing, tracked it down and while
learning about testing, discovered the usefullness of it.