Publish `minigrep` to crates.io
Despite being practically useless, minigrep is my first crate, my first
rust package. Not wanting to be that guy, I made minigrep atleast
functional, not competently though, therefore it doesn't fall in the
category of garbage in the official rust registry (crates.io) not that
they categorize in this manner.